The views have been secured for just admin

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Product, Category
 from .forms import ProductForm
@@ -57,6 +58,7 @@ def all_products(request):
     return render(request, 'products/products.html', context)
 
 
+@login_required
 def product_detail(request, product_id):
     '''
     A view to return the products details
@@ -71,10 +73,15 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     '''
     Add a product as admin to the store
     '''
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can add products')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -94,10 +101,15 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     '''
     Edit a product found in store
     '''
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can edit products')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -124,6 +136,10 @@ def delete_product(request, product_id):
     '''
     Deletes the product from the shop
     '''
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can delete products')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product has been deleted!')
